full sync when namespace labels change
